rips lots of code out of profilebuilder and moves it to nav
