--- packages/cms/src/components/interface/CtxMenu.css
@@ -1,1 +1,0 @@
-@import "../../css/mixins.css";
