--- packages/cms/src/components/interface/SidebarTree.css
@@ -1,189 +1,0 @@
-@import "../../css/mixins.css";
-
-/* --------------------------------------- */
-/* TREE */
-/* --------------------------------------- */
-
-/* "border" */
-.cms-tree > [class*="tree-node-list"] > [class*="tree-node"]:not(:last-of-type) {
-  box-shadow: 0 1px 0 var(--light-2);
-
-  @mixin dark-mode {
-    box-shadow: 0 1px 0 var(--dark-3);
-
-  }
-}
-
-.cms-tree [class*="tree-node-content-0"] {
-  min-height: 2.5rem;
-  padding-left: 0.25rem;
-}
-
-.cms-tree .bp3-tree-node {
-  cursor: pointer; /* it's not a link, but probably should be */
-
-  /* -------------------------- */
-  /* base node styles */
-  /* -------------------------- */
-  & .bp3-tree-node-content {
-    @mixin font-xs;
-    /* sizing */
-    height: auto; /* override blueprint */
-    line-height: 1.25;
-    padding-top: 0.375rem;
-    padding-bottom: 0.375rem;
-    /* alignment */
-    align-items: center; /* vertically center content */
-    transition:
-      color 0.1s ease-out,
-      background-color 0.1s ease-out;
-
-    /* reset blueprint theming */
-    &:hover {
-      background-color: transparent;
-    }
-
-    /* fudge text alignment */
-    & .bp3-tree-node-label {
-      top: -1px;
-    }
-
-    /* fudge chevron sizing and alignment */
-    & [class*="chevron"] {
-      min-width: 1.125rem;
-      height: 28px;
-      left: -0.125rem;
-      margin-left: -0.375rem;
-      margin-right: -8px;
-      top: 1px;
-
-      &[class*="open"] {
-        left: -0.2rem;
-        top: 0;
-      }
-    }
-
-    /* chevron, cog icons */
-    & [class*="bp3-icon"] {
-      /* typograhpy */
-      @mixin font-xs;
-      line-height: 0.5rem;
-      /* sizing */
-      min-width: 0; /* override blueprint */
-      padding: 0.25rem 0.5rem;
-      cursor: default; /* doesn't act like a link */
-      /* transitions */
-      transition:
-        color 0.1s ease-out,
-        transform 0.2s ease-out;
-    }
-
-    /* offset nonexistant icon ¯\_(ツ)_/¯ */
-    & .bp3-tree-node-caret-none {
-      margin-right: -30px;
-    }
-
-    /* settings "button" */
-    & .bp3-tree-node-secondary-label {
-      padding: 0; /* override blueprint */
-      margin-right: -0.275rem; /* offset icon padding (fudged) */
-
-      /* "button" transition */
-      & .bp3-icon-icon:hover {
-        transform: scale(1.25) rotate(-30deg);
-      }
-    }
-
-    /* -------------------------- */
-    /* nested nodes
-    /* -------------------------- */
-    &.bp3-tree-node-content-1 {
-      padding-left: 0.75rem;
-      font-weight: 300;
-      background-color: var(--light-2);
-      min-height: 1.875rem;
-
-      @mixin dark-mode {
-        background-color: var(--dark-3);
-      }
-
-      /* interactions */
-      &:hover {
-        background-color: color(var(--white) a(0.5));
-
-        @mixin dark-mode {
-          background-color: color(var(--black) a(0.5));
-        }
-      }
-    }
-  }
-
-  /* expanded tree background */
-  & .bp3-collapse-body:before {
-    @mixin absolute-expand;
-    content: "";
-    display: block;
-    background-color: var(--light-1);
-
-    @mixin dark-mode {
-      background-color: color(var(--black) a(0.5));
-    }
-  }
-
-  /* blueprint theming overrides */
-  & .bp3-tree-node-content:hover,
-  & .bp3-tree-node-content .bp3-tree-node-caret:hover,
-  & .bp3-tree-node-content .bp3-tree-node-cog:hover {
-    color: var(--accent-black);
-
-    @mixin dark-mode {
-      color: var(--accent-light);
-    }
-  }
-
-  /* selected state theme overrides */
-  &.bp3-tree-node-selected > .bp3-tree-node-content {
-    color: var(--accent-black);
-    &, &:hover {
-      background-color: var(--white);
-    }
-
-    /* keep icon visible on hover */
-    & [class*="bp3-icon"],
-    & [class*="bp3-icon"]:hover {
-      color: var(--dark-2);
-
-      @mixin dark-mode {
-        color: var(--light-3);
-      }
-    }
-
-    @mixin dark-mode {
-      color: var(--accent-light);
-      &, &:hover {
-        background-color: var(--black);
-      }
-
-      /* keep icon visible on hover */
-      & > .bp3-tree-node-caret:hover {
-        color: var(--accent-light);
-      }
-    }
-  }
-}
-
-[class*="tree-node"]:not(.bp3-tree-node-selected) > * > .bp3-tree-node-icon {
-  transform: scale(0.875);
-}
-
-.bp3-tree-node-icon {
-  margin-right: 0 !important;
-}
-
-.grouping-node ~ *:not(.grouping-node) > [class*="tree-node-content"] {
-  padding: 0 5px 0 2.125rem !important;
-
-  & > [class*="tree-node-label"] {
-    @mixin font-xxs;
-  }
-}
